chore(Account Creation): Move files directories
