+++ app/src/androidTest/java/com/android/unio/ui/authentication/OverlayCardTest.kt
@@ -7,8 +7,8 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import com.android.unio.model.user.Interest
+import com.android.unio.ui.authentication.overlay.InterestOverlay
 import junit.framework.TestCase.assertEquals
-import kotlinx.coroutines.flow.MutableStateFlow
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
